# Fix pandas FutureWarning and enhance wranglers performance