--- a/nautilus_trader/model/data.pxd
+++ b/nautilus_trader/model/data.pxd
@@ -367,6 +367,19 @@ cdef class QuoteTick(Data):
     )
 
     @staticmethod
+    cdef list[QuoteTick] from_raw_arrays_to_list_c(
+        InstrumentId instrument_id,
+        uint8_t price_prec,
+        uint8_t size_prec,
+        int64_t[:] bid_prices_raw,
+        int64_t[:] ask_prices_raw,
+        uint64_t[:] bid_sizes_raw,
+        uint64_t[:] ask_sizes_raw,
+        uint64_t[:] ts_events,
+        uint64_t[:] ts_inits,
+    )
+
+    @staticmethod
     cdef QuoteTick from_mem_c(QuoteTick_t mem)
 
     @staticmethod
@@ -404,6 +417,19 @@ cdef class TradeTick(Data):
         TradeId trade_id,
         uint64_t ts_event,
         uint64_t ts_init,
+    )
+
+    @staticmethod
+    cdef list[TradeTick] from_raw_arrays_to_list_c(
+        InstrumentId instrument_id,
+        uint8_t price_prec,
+        uint8_t size_prec,
+        int64_t[:] prices_raw,
+        uint64_t[:] sizes_raw,
+        uint8_t[:] aggressor_sides,
+        list[str] trade_ids,
+        uint64_t[:] ts_events,
+        uint64_t[:] ts_inits,
     )
 
     @staticmethod