# Pull Request

- Fixed pandas FutureWarning for `TradeTickDataWrangler.process_bar_data`
- Implemented `prepare_event_and_init_timestamps` which prepares `ts_event` and `ts_init` directly from `DatetimeIndex` which is already `ns`. This has improved the performance significantly, less than half time.
- Optional `sort_data` flag for `process_bar_data` with default to True

Following is the results on my dataset:
- `TradeTickDataWrangler` 1970563 rows, processed in 7s (previously 32s)
- `QuoteTickDataWrangler` 6822199 rows, processed in 13s (previously 112s)