@rsmb7z 2024-04-12
> Hey @rsmb7z
> 
> The changes look good - I see the PR is in draft mode, did you still intend to add more changes?

Hi @cjdsellers 
Yes figured out further performance improvements. Will change to 'Ready' once pushed.