@rsmb7z 2024-04-13
Hi @cjdsellers 
The performance have further improved as follows.
- 1970563 bars loaded as TradeTick in **7.00s** (from 32s)
- 1970563 bars loaded as QuoteTick in **12.89s** (from 112s)

I have also added optional flag `sort_data` set default to True for keeping current behavior. This helps further improve because `DataEngine` would be sorting eventually as well.